Move IntoLinSrgba trait into a new color::conv module
This sets a more generalised precedent for creating conversion traits
that fill the gaps left by the palette crate. This in turn allows
modules like `mesh` and `frame` to not depend on the `draw` module which
is a higher-level abstraction.
